+++ VpnHood.Server.App.Net/AppSettings.cs
@@ -1,8 +1,8 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 using System.Text.Json.Serialization;
 using VpnHood.Server.Access.Configurations;
-using VpnHood.Server.Access.Managers.File;
-using VpnHood.Server.Access.Managers.Http;
+using VpnHood.Server.Access.Managers.FileAccessManagers;
+using VpnHood.Server.Access.Managers.HttpAccessManagers;
 
 namespace VpnHood.Server.App;